Add whitelisting for webops team
This is the quick and dirty way of ensuring only webops has admin
access to the new AWS account for now.  It should also be able to
authorise lower-level roles for other departmental users.  This will be
for a later refactor, though.
