+++ rules/map-github-user-and-group-to-aws.json
@@ -1,5 +1,5 @@
 {
     "enabled": true,
-    "order": 2,
+    "order": 3,
     "stage": "login_success"
   }
