+++ rules/whitelist-github-teams.json
